requirements.txt
pandas==0.25.1
numpy==1.16.5
luigi==2.8.3
paceutils.egg==info
geopy==1.20.0
